Refactor to use custom hook
